[FLINK-7876] Merge TaskExecutorMetricsInitializer and MetricUtils
This commit removes the TaskExecutorMetricsInitializer and moves its methods
to MetricUtils.
