+++ flink-runtime/src/main/java/org/apache/flink/runtime/metrics/util/MetricUtils.java
@@ -22,23 +22,27 @@ import org.apache.flink.metrics.Gauge;
 import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.runtime.io.network.NetworkEnvironment;
 import org.apache.flink.runtime.metrics.MetricRegistry;
+import org.apache.flink.runtime.metrics.groups.JobManagerMetricGroup;
 import org.apache.flink.runtime.metrics.groups.TaskManagerMetricGroup;
-import org.apache.flink.runtime.taskexecutor.utils.TaskExecutorMetricsInitializer;
 import org.apache.flink.runtime.taskmanager.TaskManagerLocation;
+import org.apache.flink.util.Preconditions;
 
-import org.apache.commons.lang3.text.WordUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.lang.management.BufferPoolMXBean;
+import javax.management.AttributeNotFoundException;
+import javax.management.InstanceNotFoundException;
+import javax.management.MBeanException;
+import javax.management.MBeanServer;
+import javax.management.MalformedObjectNameException;
+import javax.management.ObjectName;
+import javax.management.ReflectionException;
+
 import java.lang.management.ClassLoadingMXBean;
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
 import java.lang.management.MemoryMXBean;
-import java.lang.management.OperatingSystemMXBean;
 import java.lang.management.ThreadMXBean;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.List;
 
 /**
@@ -51,6 +55,21 @@ public class MetricUtils {
 	private MetricUtils() {
 	}
 
+	public static JobManagerMetricGroup instantiateJobManagerMetricGroup(
+			final MetricRegistry metricRegistry,
+			final String hostname) {
+		final JobManagerMetricGroup jobManagerMetricGroup = new JobManagerMetricGroup(
+			metricRegistry,
+			hostname);
+
+		MetricGroup statusGroup = jobManagerMetricGroup.addGroup(METRIC_GROUP_STATUS_NAME);
+
+		// initialize the JM metrics
+		instantiateStatusMetrics(statusGroup);
+
+		return jobManagerMetricGroup;
+	}
+
 	public static TaskManagerMetricGroup instantiateTaskManagerMetricGroup(
 			MetricRegistry metricRegistry,
 			TaskManagerLocation taskManagerLocation,
@@ -60,41 +79,18 @@ public class MetricUtils {
 			taskManagerLocation.getHostname(),
 			taskManagerLocation.getResourceID().toString());
 
+		MetricGroup statusGroup = taskManagerMetricGroup.addGroup(METRIC_GROUP_STATUS_NAME);
+
 		// Initialize the TM metrics
-		TaskExecutorMetricsInitializer.instantiateStatusMetrics(taskManagerMetricGroup, network);
+		instantiateStatusMetrics(statusGroup);
+		instantiateNetworkMetrics(statusGroup, network);
 
 		return taskManagerMetricGroup;
 	}
 
-	public static void instantiateNetworkMetrics(
-		MetricGroup metrics,
-		final NetworkEnvironment network) {
-		MetricGroup status = metrics.addGroup(METRIC_GROUP_STATUS_NAME);
-
-		MetricGroup networkGroup = status
-			.addGroup("Network");
-
-		networkGroup.gauge("TotalMemorySegments", new Gauge<Integer>() {
-			@Override
-			public Integer getValue() {
-				return network.getNetworkBufferPool().getTotalNumberOfMemorySegments();
-			}
-		});
-		networkGroup.gauge("AvailableMemorySegments", new Gauge<Integer>() {
-			@Override
-			public Integer getValue() {
-				return network.getNetworkBufferPool().getNumberOfAvailableMemorySegments();
-			}
-		});
-	}
-
 	public static void instantiateStatusMetrics(
-		MetricGroup metrics) {
-		MetricGroup status = metrics
-			.addGroup(METRIC_GROUP_STATUS_NAME);
-
-		MetricGroup jvm = status
-			.addGroup("JVM");
+			MetricGroup metricGroup) {
+		MetricGroup jvm = metricGroup.addGroup("JVM");
 
 		instantiateClassLoaderMetrics(jvm.addGroup("ClassLoader"));
 		instantiateGarbageCollectorMetrics(jvm.addGroup("GarbageCollector"));
@@ -103,16 +99,35 @@ public class MetricUtils {
 		instantiateCPUMetrics(jvm.addGroup("CPU"));
 	}
 
+	private static void instantiateNetworkMetrics(
+		MetricGroup metrics,
+		final NetworkEnvironment network) {
+		metrics.<Long, Gauge<Long>>gauge("TotalMemorySegments", new Gauge<Long> () {
+			@Override
+			public Long getValue() {
+				return (long) network.getNetworkBufferPool().getTotalNumberOfMemorySegments();
+			}
+		});
+
+		metrics.<Long, Gauge<Long>>gauge("AvailableMemorySegments", new Gauge<Long> () {
+			@Override
+			public Long getValue() {
+				return (long) network.getNetworkBufferPool().getNumberOfAvailableMemorySegments();
+			}
+		});
+	}
+
 	private static void instantiateClassLoaderMetrics(MetricGroup metrics) {
 		final ClassLoadingMXBean mxBean = ManagementFactory.getClassLoadingMXBean();
 
-		metrics.gauge("ClassesLoaded", new Gauge<Long>() {
+		metrics.<Long, Gauge<Long>>gauge("ClassesLoaded", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getTotalLoadedClassCount();
 			}
 		});
-		metrics.gauge("ClassesUnloaded", new Gauge<Long>() {
+
+		metrics.<Long, Gauge<Long>>gauge("ClassesUnloaded", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getUnloadedClassCount();
@@ -123,15 +138,17 @@ public class MetricUtils {
 	private static void instantiateGarbageCollectorMetrics(MetricGroup metrics) {
 		List<GarbageCollectorMXBean> garbageCollectors = ManagementFactory.getGarbageCollectorMXBeans();
 
-		for (final GarbageCollectorMXBean garbageCollector : garbageCollectors) {
+		for (final GarbageCollectorMXBean garbageCollector: garbageCollectors) {
 			MetricGroup gcGroup = metrics.addGroup(garbageCollector.getName());
-			gcGroup.gauge("Count", new Gauge<Long>() {
+
+			gcGroup.<Long, Gauge<Long>>gauge("Count", new Gauge<Long> () {
 				@Override
 				public Long getValue() {
 					return garbageCollector.getCollectionCount();
 				}
 			});
-			gcGroup.gauge("Time", new Gauge<Long>() {
+
+			gcGroup.<Long, Gauge<Long>>gauge("Time", new Gauge<Long> () {
 				@Override
 				public Long getValue() {
 					return garbageCollector.getCollectionTime();
@@ -142,20 +159,22 @@ public class MetricUtils {
 
 	private static void instantiateMemoryMetrics(MetricGroup metrics) {
 		final MemoryMXBean mxBean = ManagementFactory.getMemoryMXBean();
+
 		MetricGroup heap = metrics.addGroup("Heap");
-		heap.gauge("Used", new Gauge<Long>() {
+
+		heap.<Long, Gauge<Long>>gauge("Used", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getHeapMemoryUsage().getUsed();
 			}
 		});
-		heap.gauge("Committed", new Gauge<Long>() {
+		heap.<Long, Gauge<Long>>gauge("Committed", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getHeapMemoryUsage().getCommitted();
 			}
 		});
-		heap.gauge("Max", new Gauge<Long>() {
+		heap.<Long, Gauge<Long>>gauge("Max", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getHeapMemoryUsage().getMax();
@@ -163,54 +182,61 @@ public class MetricUtils {
 		});
 
 		MetricGroup nonHeap = metrics.addGroup("NonHeap");
-		nonHeap.gauge("Used", new Gauge<Long>() {
+
+		nonHeap.<Long, Gauge<Long>>gauge("Used", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getNonHeapMemoryUsage().getUsed();
 			}
 		});
-		nonHeap.gauge("Committed", new Gauge<Long>() {
+		nonHeap.<Long, Gauge<Long>>gauge("Committed", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getNonHeapMemoryUsage().getCommitted();
 			}
 		});
-		nonHeap.gauge("Max", new Gauge<Long>() {
+		nonHeap.<Long, Gauge<Long>>gauge("Max", new Gauge<Long> () {
 			@Override
 			public Long getValue() {
 				return mxBean.getNonHeapMemoryUsage().getMax();
 			}
 		});
 
-		List<BufferPoolMXBean> bufferMxBeans = ManagementFactory.getPlatformMXBeans(BufferPoolMXBean.class);
+		final MBeanServer con = ManagementFactory.getPlatformMBeanServer();
 
-		for (final BufferPoolMXBean bufferMxBean : bufferMxBeans) {
-			MetricGroup bufferGroup = metrics.addGroup(WordUtils.capitalize(bufferMxBean.getName()));
-			bufferGroup.gauge("Count", new Gauge<Long>() {
-				@Override
-				public Long getValue() {
-					return bufferMxBean.getCount();
-				}
-			});
-			bufferGroup.gauge("MemoryUsed", new Gauge<Long>() {
-				@Override
-				public Long getValue() {
-					return bufferMxBean.getMemoryUsed();
-				}
-			});
-			bufferGroup.gauge("TotalCapacity", new Gauge<Long>() {
-				@Override
-				public Long getValue() {
-					return bufferMxBean.getTotalCapacity();
-				}
-			});
+		final String directBufferPoolName = "java.nio:type=BufferPool,name=direct";
+
+		try {
+			final ObjectName directObjectName = new ObjectName(directBufferPoolName);
+
+			MetricGroup direct = metrics.addGroup("Direct");
+
+			direct.<Long, Gauge<Long>>gauge("Count", new AttributeGauge<>(con, directObjectName, "Count", -1L));
+			direct.<Long, Gauge<Long>>gauge("MemoryUsed", new AttributeGauge<>(con, directObjectName, "MemoryUsed", -1L));
+			direct.<Long, Gauge<Long>>gauge("TotalCapacity", new AttributeGauge<>(con, directObjectName, "TotalCapacity", -1L));
+		} catch (MalformedObjectNameException e) {
+			LOG.warn("Could not create object name {}.", directBufferPoolName, e);
+		}
+
+		final String mappedBufferPoolName = "java.nio:type=BufferPool,name=mapped";
+
+		try {
+			final ObjectName mappedObjectName = new ObjectName(mappedBufferPoolName);
+
+			MetricGroup mapped = metrics.addGroup("Mapped");
+
+			mapped.<Long, Gauge<Long>>gauge("Count", new AttributeGauge<>(con, mappedObjectName, "Count", -1L));
+			mapped.<Long, Gauge<Long>>gauge("MemoryUsed", new AttributeGauge<>(con, mappedObjectName, "MemoryUsed", -1L));
+			mapped.<Long, Gauge<Long>>gauge("TotalCapacity", new AttributeGauge<>(con, mappedObjectName, "TotalCapacity", -1L));
+		} catch (MalformedObjectNameException e) {
+			LOG.warn("Could not create object name {}.", mappedBufferPoolName, e);
 		}
 	}
 
 	private static void instantiateThreadMetrics(MetricGroup metrics) {
 		final ThreadMXBean mxBean = ManagementFactory.getThreadMXBean();
 
-		metrics.gauge("Count", new Gauge<Integer>() {
+		metrics.<Integer, Gauge<Integer>>gauge("Count", new Gauge<Integer> () {
 			@Override
 			public Integer getValue() {
 				return mxBean.getThreadCount();
@@ -220,54 +246,48 @@ public class MetricUtils {
 
 	private static void instantiateCPUMetrics(MetricGroup metrics) {
 		try {
-			final OperatingSystemMXBean mxBean = ManagementFactory.getOperatingSystemMXBean();
+			final com.sun.management.OperatingSystemMXBean mxBean = (com.sun.management.OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
 
-			final Method fetchCPULoadMethod = Class.forName("com.sun.management.OperatingSystemMXBean")
-				.getMethod("getProcessCpuLoad");
-			// verify that we can invoke the method
-			fetchCPULoadMethod.invoke(mxBean);
-
-			final Method fetchCPUTimeMethod = Class.forName("com.sun.management.OperatingSystemMXBean")
-				.getMethod("getProcessCpuTime");
-			// verify that we can invoke the method
-			fetchCPUTimeMethod.invoke(mxBean);
-
-			metrics.gauge("Load", new Gauge<Double>() {
+			metrics.<Double, Gauge<Double>>gauge("Load", new Gauge<Double> () {
 				@Override
 				public Double getValue() {
-					try {
-						return (Double) fetchCPULoadMethod.invoke(mxBean);
-					} catch (IllegalAccessException | InvocationTargetException | IllegalArgumentException ignored) {
-						return -1.0;
-					}
+					return mxBean.getProcessCpuLoad();
 				}
 			});
-			metrics.gauge("Time", new Gauge<Long>() {
+			metrics.<Long, Gauge<Long>>gauge("Time", new Gauge<Long> () {
 				@Override
 				public Long getValue() {
-					try {
-						return (Long) fetchCPUTimeMethod.invoke(mxBean);
-					} catch (IllegalAccessException | InvocationTargetException | IllegalArgumentException ignored) {
-						return -1L;
-					}
+					return mxBean.getProcessCpuTime();
 				}
 			});
-		} catch (ClassNotFoundException | InvocationTargetException | SecurityException | NoSuchMethodException | IllegalArgumentException | IllegalAccessException ignored) {
+		} catch (Exception e) {
 			LOG.warn("Cannot access com.sun.management.OperatingSystemMXBean.getProcessCpuLoad()" +
-				" - CPU load metrics will not be available.");
-			// make sure that a metric still exists for the given name
-			metrics.gauge("Load", new Gauge<Double>() {
-				@Override
-				public Double getValue() {
-					return -1.0;
-				}
-			});
-			metrics.gauge("Time", new Gauge<Long>() {
-				@Override
-				public Long getValue() {
-					return -1L;
-				}
-			});
+				" - CPU load metrics will not be available.", e);
+		}
+	}
+
+	private static final class AttributeGauge<T> implements Gauge<T> {
+		private final MBeanServer server;
+		private final ObjectName objectName;
+		private final String attributeName;
+		private final T errorValue;
+
+		private AttributeGauge(MBeanServer server, ObjectName objectName, String attributeName, T errorValue) {
+			this.server = Preconditions.checkNotNull(server);
+			this.objectName = Preconditions.checkNotNull(objectName);
+			this.attributeName = Preconditions.checkNotNull(attributeName);
+			this.errorValue = errorValue;
+		}
+
+		@SuppressWarnings("unchecked")
+		@Override
+		public T getValue() {
+			try {
+				return (T) server.getAttribute(objectName, attributeName);
+			} catch (MBeanException | AttributeNotFoundException | InstanceNotFoundException | ReflectionException e) {
+				LOG.warn("Could not read attribute {}.", attributeName, e);
+				return errorValue;
+			}
 		}
 	}
 }
